Handle browser history
Previously the back and forwards buttons were broken. This change
listens to history change events and responds appropriately.

Note the function call

```
history.listen((location) => setKey(location.pathname))
```

returns an `unlisten` function, so this construction returns `unlisten`
and calls it when the component unmounts. This is quite an elegant way
of automagically ensuring this pattern.